+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # row to object
 ## revisions
 
+## 2020-04-22
+- added guid(length) and md5
+
 ## 2020-04-18
 - \#emptyCheck to change the behavour of the empty remove. Values are none, length
+- added the function replace
 
 ## 2020-04-15 0.9.3
 - n2br